Add lint docs and fix lint issues (#1)
* Add note on how to self check the linting

* Fix the lint issues

* Fix markdownlint error

From build:

docs/index.md: 1: MD041/first-line-h1 First line in file should be a
top level header [Context: "Welcome to the guided tour on ..."]

https://travis-ci.org/smartsdk/guided-tour/jobs/334686547#L451

* Revert "Add note on how to self check the linting"

The linting mentioned rules in the docs are different from the linting
tool in the build system.  Do not add more confusion with this
reference.  The original doc will be fixed later.

This reverts commit 9092d28577e670c1afe2bfb0f69f71aa1f660359.

* Fix the markdownlint repository link

The markdownlint package is installed for the CI using npm:

https://www.npmjs.com/package/markdownlint

The correct source repo is:

https://github.com/DavidAnson/markdownlint

And not another with similar name:

https://github.com/markdownlint/markdownlint